+++ src-lazy/app/app.routes.ts
@@ -2,11 +2,10 @@ import { ModuleWithProviders } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { HomeComponent } from './shared/components/home/home.component';
-import { LazyComponent } from './shared/components/lazy/lazy.component';
 
 const routes: Routes = [
-  { path: '', component: HomeComponent },
-  { path: 'lazy', component: LazyComponent }
+    { path: '', component: HomeComponent },
+    { path: 'lazy', loadChildren: 'shared/components/lazy/lazy.module#LazyModule' }
 ];
 
 export const routing: ModuleWithProviders = RouterModule.forRoot(routes);