+++ src-lazy/public/system.config.prod.js
@@ -3,11 +3,12 @@
   System.config({
     'map': {
       'bundle': 'bundle.js',
-      'vendor': 'vendor.js'
+      'vendor': 'vendor.js',
+      'polyfill': 'system.polyfill.js'
     },
     'meta': {
       'bundle': {
-        deps: []
+        deps: ['polyfill']
       }
     }
   });