+++ src-lazy/public/system.import.js
@@ -1,0 +1,8 @@
+System.import('system.config.prod.js').then(function () {
+  System.import('polyfill').then(function () {
+    Promise.all([
+      System.import('vendor'),
+      System.import('bundle')
+    ]);
+  }).catch(console.error.bind(console));
+}).catch(console.error.bind(console));