+++ src/app/app.module.ts
@@ -7,6 +7,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AppComponent }  from './app.component';
 import { routing }       from './app.routes';
 import { HomeModule } from './shared/components/home/home.module';
+import { LazyModule } from './shared/components/lazy/lazy.module';
 
 @NgModule({
 
@@ -16,6 +17,7 @@ import { HomeModule } from './shared/components/home/home.module';
                CommonModule,
                FormsModule,
                HomeModule,
+               LazyModule,
                routing],
     declarations: [ AppComponent ],
     bootstrap:    [ AppComponent ]