+++ src/public/system.import.js
@@ -1,8 +1,5 @@
 System.import('system.config.prod.js').then(function () {
-  System.import('polyfill').then(function () {
-    Promise.all([
-      System.import('vendor'),
-      System.import('bundle')
-    ]);
-  }).catch(console.error.bind(console));
+  Promise.all([
+    System.import('bundle')
+  ]);
 }).catch(console.error.bind(console));